changed process_data

+++ conv_net_sentence.py
@@ -312,8 +312,8 @@ def make_idx_data_cv(revs, mapping, cv, max_l=51, filter_h=5):
             test.append(sent)
         else:
             train.append(sent)
-    train = np.array(train, dtype="int32")
-    test = np.array(train, dtype="int32")
+    train = np.array(train)
+    test = np.array(train)
     print train.shape, test.shape
     return [train, test]
